adjusted model classes
