+++ app/src/main/java/com/arctouch/codechallenge/model/GenreResponse.kt
@@ -1,0 +1,2 @@
+package com.arctouch.codechallenge.model
+
