+++ app/src/main/java/com/arctouch/codechallenge/model/Movie.kt
@@ -1,0 +1,25 @@
+package com.arctouch.codechallenge.model
+
+import com.google.gson.annotations.SerializedName
+
+data class GenreResponse(val genres: List<Genre>)
+
+data class Genre(val id: Int, val name: String)
+
+data class UpcomingMoviesResponse(
+        val page: Int,
+        val results: List<Movie>,
+        @SerializedName("total_pages") val totalPages: Int,
+        @SerializedName("total_results") val totalResults: Int
+)
+
+data class Movie(
+    val id: Int,
+    val title: String,
+    val overview: String?,
+    val genres: List<Genre>?,
+    @SerializedName("genre_ids") val genreIds: List<Int>?,
+    @SerializedName("poster_path") val posterPath: String?,
+    @SerializedName("backdrop_path") val backdropPath: String?,
+    @SerializedName("release_date") val releaseDate: String?
+)
